+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change Log
 
-All notable changes to the "rpkg-list" extension will be documented in this file.
+All notable changes to the "rpkgs-list" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 
-- Initial release
+- [x] Access R library paths
+- [x] Read `DESCRIPTION` files and extract variables
+- [x] Create a data model
+- [x] Register commands/UX views
+- [ ] Get data to show up in view panel